backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.6
pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
googlemaps==4.10.0
